labs/plot_data/requirements.txt
# numpy==1.19.5
# scipy==1.5.2
# pyqtgraph==0.12.2
# pyserial==3.5
# PyOpenGL==3.1.5
# spyder==5.1.1
# PyQt5==5.15.4

numpy
scipy
pyqtgraph
pyserial
PyOpenGL
spyder
PyQt5